requirements.txt
streamlit
requests
google-generativeai
folium
streamlit_folium
fpdf
geopy